Alert Database Size Calculation Error
A situation can occur where the number in mongo is so big that they decide to encapsulate it inside of NumberLong("...."). This will break the script. I have added a sed command to filter out that if it should ever exist and also modfiied the megabyte size to be 1 million.